+++ lib/gecko/helpers/association_helper.rb
@@ -7,7 +7,6 @@ module Gecko
       end
 
       module ClassMethods
-
         # Set up a belongs_to relationship between two classes based on a
         # JSON key of {class_name}_id.
         #
@@ -48,8 +47,10 @@ module Gecko
         # @param [Symbol] model_name
         # @param [#to_hash] options options for setting up the relationship
         # @option options [String] :class_name Override the default class name
-        # @option options [String] :readonly (true) Whether to serialize this
+        # @option options [Boolean] :readonly (true) Whether to serialize this
         #   attribute
+        # @option options [Boolean] :embedded (false) Whether this relation should
+        #   persisted inside it's parent record
         #
         # @return [undefined]
         #
@@ -61,15 +62,114 @@ module Gecko
           readonly    = options.key?(:readonly) ? options[:readonly] : true
 
           define_method association_name do
-            ids = self.attributes[foreign_key]
-            if ids.any?
-              @client.adapter_for(class_name).find_many(ids)
-            else
-              []
+            collection_proxies[association_name] ||= begin
+              ids = self.attributes[foreign_key]
+              if ids.any?
+                collection = @client.adapter_for(class_name).find_many(ids)
+              else
+                collection = []
+              end
+
+              build_collection_proxy(collection, {
+                embedded:         options[:embedded],
+                class_name:       class_name,
+                association_name: association_name
+              })
             end
           end
           attribute foreign_key, Array[Integer], readonly: readonly
         end
+      end
+
+    private
+
+      # Stores a reference to all of the child collection proxies on the model
+      #
+      # @return [Hash]
+      #
+      # @api private
+      def collection_proxies
+        @collection_proxies ||= {}
+      end
+
+      # Setup a child collection proxy on first instantation
+      #
+      # @return [Hash]
+      #
+      # @api private
+      def build_collection_proxy(target, association_name:, class_name:, embedded:)
+        CollectionProxy.new({
+          parent:           self,
+          target:           target,
+          embedded:         embedded,
+          class_name:       class_name,
+          association_name: association_name
+        })
+      end
+    end
+
+    # Provides a convenient wrapper for a collection of child records.
+    # Exposes both an Enumerable interface as well as the ability
+    # to create new child records
+    class CollectionProxy
+      include Enumerable
+      delegate :each, :empty?, to: :@target
+
+      attr_reader :association_name
+
+      # Setup the child collection proxy
+      #
+      # @return [Hash]
+      #
+      # @api private
+      def initialize(parent:, association_name:, class_name:, target:, embedded:)
+        @parent           = parent
+        @target           = target
+        @embedded         = embedded
+        @class_name       = class_name
+        @association_name = association_name
+      end
+
+      # Build a new child object inside the collection
+      #
+      # @example
+      #   item = order.order_line_items.build(variant_id: 1234, quantiy: 12.5, price: 13.45)
+      #   order.order_line_items.include?(item) #=> true
+      #
+      # @param [#to_hash] attributes for the child record
+      #
+      # @return [Gecko::Record::Base]
+      #
+      # @api public
+      def build(attributes)
+        if @parent.persisted?
+          parent_foreign_key = @parent.class.demodulized_name.foreign_key.to_sym
+          attributes[parent_foreign_key] = @parent.id
+        end
+
+        record = client.adapter_for(@class_name).build(attributes)
+        @target << record
+        record
+      end
+
+      # Should this collection of records be serialized inside it's parent object
+      #
+      # @return [Boolean]
+      #
+      # @api private
+      def embed_records?
+        !!@embedded
+      end
+
+    private
+
+      # Access the Gecko Client object
+      #
+      # @return [Gecko::Client]
+      #
+      # @api private
+      def client
+        @parent.instance_variable_get(:@client)
       end
     end
   end